#3 Merge Setting Pages

+++ pages/SettingsPage/SettingsPage.tsx
@@ -1,24 +1,29 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { createBox, createText } from '@shopify/restyle';
 import theme, { Theme } from '../../utils/theme';
 import { useDispatch } from "react-redux";
 import { AppDispatch } from "../../redux/store";
 import { selectUsername, selectToken, addUser, renameUser, deleteUser } from "../../redux/slice/userSlice";
 import { useSelector } from "react-redux";
-import IconButton from "../../components/ui/inputs/IconButton";
 import TextButton from "../../components/ui/inputs/TextButton";
 import { NavigationScreenProp } from "react-navigation";
 import { UsernameInput } from "../../components/ui/inputs/UsernameInput";
 import Ionicons from 'react-native-vector-icons/Ionicons'
+import { Restriction } from "../../utils/dataTypes";
+import { getRestrictions, setRestrictions } from "../../utils/axios/userManagementCalls";
+import { ScrollView } from "react-native";
+import { RestrictionButton } from "../../components/ui/inputs/RestrictionButton";
+
 
 const Text = createText<Theme>();
 const Box = createBox<Theme>();
 
-export type UserPageProps = {
+export type SettingsPageProps = {
   navigation: NavigationScreenProp<any, any>
 };
 
-const UserPage = ({ navigation }: UserPageProps) => {
+const SettingsPage = ({ navigation }: SettingsPageProps) => {
+  // User Management
   const username = useSelector(selectUsername)
   const token = useSelector(selectToken)
   const [currentUsername, setCurrentUsername] = useState("")
@@ -27,19 +32,44 @@ const UserPage = ({ navigation }: UserPageProps) => {
   const dispatch = useDispatch<AppDispatch>()
 
   const changeUser = (newUsername: string) => {
-    if (username === "") {
-      dispatch(addUser(newUsername))
-    } else {
-      dispatch(renameUser(newUsername))
-    }
+    dispatch(renameUser(newUsername))
   }
 
   const deleteUsername = () => {
     /*
-    if (username != "dev" && token.startsWith('T')) {
+    if (username != "dev" && username != "Guest" && token.startsWith('T')) {
       dispatch(deleteUser(username))
     }
     */
+  }
+
+  // Restriction Management
+
+  const [currentRestrictions, setCurrentRestrictions] = useState([] as Restriction[])
+
+  // Fetch available restrictions
+  useEffect(() => {
+    getRestrictions().then(
+      (availableRestrictions: Restriction[]) => {
+        setCurrentRestrictions(availableRestrictions)
+      }
+    )
+  }, [])
+
+  const setNewRestriction = async (restriction: string) => {
+    const success: boolean = await setRestrictions(restriction)
+    if (success) {
+      const current = currentRestrictions.map((x) => x)
+      for (const restr of current) {
+        if (restr.active) {
+          restr.active = false
+        }
+        if (!restr.active && restr.name === restriction) {
+          restr.active = true
+        }
+      }
+      setCurrentRestrictions(current)
+    }
   }
 
   return (
@@ -54,7 +84,7 @@ const UserPage = ({ navigation }: UserPageProps) => {
         </Box>
       </Box>
 
-      <Box marginTop={"l"}>
+      <Box marginTop={"m"}>
         <UsernameInput
           clicked={clicked}
           setClicked={setClicked}
@@ -65,7 +95,7 @@ const UserPage = ({ navigation }: UserPageProps) => {
         />
       </Box>
 
-      <Box marginTop={"l"}>
+      <Box marginTop={"m"} marginBottom={"m"}>
         <TextButton
           onPress={deleteUsername}
           icon={"close-circle-outline"}
@@ -74,18 +104,28 @@ const UserPage = ({ navigation }: UserPageProps) => {
           color={theme.colors.black}
         />
       </Box>
-      <Box flexGrow={1} />
 
-      <Box alignItems={"center"}>
-        <IconButton
-          onPress={() => navigation.navigate("Restrictions")}
-          icon={'chevron-forward-circle-outline'}
-          size={60}
-          color={theme.colors.inactiveButtonColor} />
+      <Box marginTop={"m"} marginBottom={"m"} alignItems="center">
+        <Text variant="subsubheader">Choose your Diet...</Text>
       </Box>
+
+      <ScrollView showsVerticalScrollIndicator={false}>
+        {
+          currentRestrictions.map((elem: Restriction) => {
+            return (
+              <RestrictionButton
+                key={elem.name}
+                restriction={elem}
+                setNewRestriction={setNewRestriction}
+              />
+            )
+          })
+        }
+      </ScrollView>
+      <Box />
 
     </Box>
   );
 }
 
-export default UserPage; 
+export default SettingsPage;
